+++ web_portal/src/services/reservationService.js
@@ -1,4 +1,4 @@
-import { ref, update, get, serverTimestamp, increment } from "firebase/database";
+import { ref, update, get, serverTimestamp, increment, push } from "firebase/database";
 import { db } from "../firebase";
 
 /**